refactor: BaseBall 객체 제거
